splitting admin bookmark controller
