+++ shoebox/app/views/admin.scala.html
@@ -36,7 +36,7 @@
 		  <div class="navbar-inner">
 		    <ul class="nav">
 		      <li><a href="@com.keepit.controllers.admin.routes.AdminDashboardController.index">Dashboard</a>
-		      <li><a href="@com.keepit.controllers.routes.BookmarksController.bookmarksView(0)">Keeps</a>
+		      <li><a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(0)">Keeps</a>
 		      <li><a href="@com.keepit.controllers.routes.UserController.usersView">Users</a>
 		      <li><a href="@com.keepit.controllers.routes.SocialUserController.socialUsersView(0)">Social Users</a>
 		      <li><a href="@com.keepit.controllers.routes.CommentController.followsView">Follows</a>
