+++ shoebox/app/views/adminHelper/bookmarkDisplay.scala.html
@@ -1,6 +1,6 @@
 @(bookmark: Bookmark)
 
-<a href="@com.keepit.controllers.routes.BookmarksController.edit(bookmark.id.get)">
+<a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.edit(bookmark.id.get)">
 @adminHelper.bookmarkBadge(bookmark)
 @{if (bookmark.title.size > 70) bookmark.title.substring(0, 70) + "..." else bookmark.title}
 </a>
