+++ shoebox/app/views/bookmark.scala.html
@@ -46,7 +46,7 @@
      });
    </script>
    <p>
-     @helper.form(action = com.keepit.controllers.routes.BookmarksController.delete(bookmark.id.get)) {
+     @helper.form(action = com.keepit.controllers.admin.routes.AdminBookmarksController.delete(bookmark.id.get)) {
        <input type="submit" value="Delete">
      }
    </p>
