+++ shoebox/app/views/bookmarks.scala.html
@@ -6,7 +6,7 @@
     <ul>
       <li>
         @if(page > 0) {
-          <a href="@com.keepit.controllers.routes.BookmarksController.bookmarksView(page - 1)">
+          <a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(page - 1)">
         } else {
          <a href="#">
         }
@@ -16,17 +16,17 @@
       @for(i <- 0 to (pageCount - 1)) {
         @if(i == page) {
 	        <li class="disabled">
-	          <a href="@com.keepit.controllers.routes.BookmarksController.bookmarksView(i)">@{i + 1}</a>
+	          <a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(i)">@{i + 1}</a>
 	        </li>
         } else {
 	        <li class="active">
-	          <a href="@com.keepit.controllers.routes.BookmarksController.bookmarksView(i)">@{i + 1}</a>
+	          <a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(i)">@{i + 1}</a>
 	        </li>
 	      }
       }
       <li>
         @if(bookmarksAndUsers.size > 0) {
-          <a href="@com.keepit.controllers.routes.BookmarksController.bookmarksView(page + 1)">
+          <a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(page + 1)">
         } else {
          <a href="#">
         }
